processor.CMDIRecordProcessor renamed to .CMDIRecordImporter. Also added some javadoc.
